chore: deprecate memory_diff_proposer.py — archive implementation to tools/deprecated_tools
